scripts/ipkg-build: observe package provided conffiles
If a package provided a directory in conffiles, this was ignored, and
replaced with an explicit list of files in that directory.  This
prevents an application from declaring a directory that _will_ contain
files that must be preserved, and behaves differently to being able to
list directories explicitly in /etc/sysupgrade.conf.  There doesn't
appear to be any reason for this "resolution" process, so simply drop
it, and use the conffiles as provided by the package.

Signed-off-by: Karl Palsson <karlp@etactica.com>