ath79: eg200: mark the modbus light as oneshot in the dts
This ensures that the trigger file is available from boot.  Previously,
there was a window where the openwrt led system was (re)setting triggers
based on config, and the trigger would be "none" for a short time,
resulting in mrelay starting up but not having access to the trigger
file.  This also paves the way forward for simply dropping the openwrt
configuration of this, however the properties for setting the
delay_on/off values aren't in all kernels yet.

Signed-off-by: Karl Palsson <karlp@etactica.com>